Updated Main Scene
Added an Empty GameObject to the Hierarchy and renamed it Graph.
Applied the Graph script component to the Graph game object.
Provided the necessary Inspector references for the Graph script component.
Created a Cube game object and renamed it to Point.
Applied the Point material to the Point game object.
Applied the Point shader to the Point material.
Created a prefab of the Point game object.